refactor: folders hierarchy module-component
