+++ src/app/app-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomePageComponent } from './components/home-page/home-page.component';
+import { HomePageComponent } from './modules/shared/home-page/home-page.component';
 
 const routes: Routes = [
   {
