+++ src/app/app.module.ts
@@ -6,12 +6,12 @@ import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { SharedModule } from './modules/shared/shared.module';
-import { HomePageComponent } from './components/home-page/home-page.component';
+import { HomePageComponent } from './modules/shared/home-page/home-page.component';
 import { initializeApp,provideFirebaseApp } from '@angular/fire/app';
 import { environment } from '../environments/environment';
 import { provideAuth,getAuth } from '@angular/fire/auth';
 import { provideDatabase,getDatabase } from '@angular/fire/database';
-import { LoginPageComponent } from './components/login-page/login-page.component';
+import { LoginPageComponent } from './modules/user/login-page/login-page.component';
 
 @NgModule({
   declarations: [
