+++ src/app/modules/user/user-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { LoginPageComponent } from 'src/app/components/login-page/login-page.component';
+import { LoginPageComponent } from 'src/app/modules/user/login-page/login-page.component';
 
 const routes: Routes = [
   { path: '', component: LoginPageComponent },
